% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_requests.R
\name{create_request}
\alias{create_request}
\title{Create a HTTP request}
\usage{
create_request(
  url_type = "standard",
  form_id = NULL,
  request_type = "form_list",
  limit = NULL
)
}
\arguments{
\item{url_type}{A string value: "standard", "EU", "HIPAA", or "custom". "standard" is the default value.}

\item{form_id}{A string value. Default value is NULL. This is the series of digits that can be found at the end of the form URL. If form IDs are unknown, create a HTTP request of request_type = "form_list". This will return a list of forms and includes the form ID for each.}

\item{request_type}{A string value: "form_list", "form", or "submissions". Default value is "form_list". "form_list" returns all user form data up to the specified or default limit. "form" returns properties of a specified form. "submissions" returns submission data of a specified form.}

\item{limit}{A numeric value between 1-1000. Default value is NULL. JotForm by default will pull data on up to 20 content types, i.e. data on 20 forms, 20 submissions, etc. Use this field to increase or decrease data pulled.}
}
\value{
An HTTP response object. Contains metadata, headers, and body of the response.
}
\description{
This function creates a HTTP request to send to JotForm.
}
