% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_to_df}
\alias{parse_to_df}
\title{Parse response and return as data frame}
\usage{
parse_to_df(request, type, pivot)
}
\arguments{
\item{request}{A HTTP response object created by \code{\link[=create_request]{create_request()}}.}

\item{type}{A string value: "form_list", "form", or "submissions". "form_list" returns a list of the user's forms. "form" returns properties of a specified form. "submissions" returns submission data of a specified form.}

\item{pivot}{A string value used only for type = "submissions". Determines shape of final dataframe, options are "wide" or "long".}
}
\value{
A data frame containing the data type specified in the "type" argument.
}
\description{
This function takes the JotForm HTTP response object, extracts the body as a parsed JSON, and returns a data frame of the specified JotForm data.
}
