% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_key}
\alias{set_key}
\title{Set JotForm API Key}
\usage{
set_key(key)
}
\arguments{
\item{key}{A string value of the JotForm API key.}
}
\description{
This function stores the JotForm API Key in .Rprofile.
}
